+++ src/main/java/racingcar/Car.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import java.util.ArrayList;
+
 public class Car {
     private String name;
     private int position;
@@ -20,7 +22,35 @@ public class Car {
         return position;
     }
 
-    public void move(){
+    public void moveOnePoint(){
         this.position += 1;
     }
+
+    public static ArrayList<Car> decideWinner(ArrayList<Car> car){
+        ArrayList<Car> winners = new ArrayList<Car>();
+        int maxPosition = car.get(0).position;
+        for(Car selectedCar : car){
+            maxPosition = searchMaxPosition(maxPosition, selectedCar);
+        }
+
+        for(Car selectedCar : car){
+            checkPositionPerCar(maxPosition, selectedCar, winners);
+        }
+
+        return winners;
+    }
+
+    public static int searchMaxPosition(int maxPosition, Car car){
+        if(car.position >= maxPosition){
+            return car.position;
+        }
+        return maxPosition;
+    }
+
+    public static void checkPositionPerCar(int maxPosition, Car car, ArrayList<Car> winners){
+        if(car.position == maxPosition){
+            winners.add(car);
+        }
+    }
+
 }
